+++ src/Manager/CallbackJobManager.php
@@ -29,7 +29,7 @@ final class CallbackJobManager implements JobManager
 		}
 	}
 
-	public function getPair($id): ?array
+	public function getScheduledJob($id): ?array
 	{
 		$job = $this->jobs[$id] ?? null;
 
@@ -43,7 +43,7 @@ final class CallbackJobManager implements JobManager
 		];
 	}
 
-	public function getPairs(): array
+	public function getScheduledJobs(): array
 	{
 		$pairs = [];
 		foreach ($this->jobs as $id => $job) {