+++ CHANGELOG.md
@@ -14,3 +14,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 	- `getPairs()` -> `getScheduledJobs()`
 - `Scheduler`
 	- `getJobs()` -> `getScheduledJobs()`
+- `RunSummary`
+	- `getJobs()` -> `getJobSummaries()`